% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_score}
\alias{df_score}
\title{Annual parameter scores (dataframe)}
\format{
A dataframe with 15-17 columns:
 \describe{
   \item{Year}{Year}
   \item{Site_Name}{Site name}
   \item{Site_ID}{Site ID}
   \item{Town}{Town}
   \item{Watereshed}{Watershed}
   \item{Group}{Site group}
   \item{Depth}{Depth category}
   \item{Parameter}{Parameter}
   \item{Unit}{Result unit}
   \item{score_typ}{How annual score is calculated. Values: Minimum, Maximum,
   Average, Geomean, Median, 90th percentile, NA}
   \item{score_num}{Numeric score}
   \item{score_str}{Score. Values: Meets Criteria, Does Not Meet Criteria,
   Excellent, Good, Fair, Poor, No Threshold Established, No Data Available}
   \item{Longitude}{Longitude}
   \item{Latitude}{Latitude}
   \item{popup_loc}{Popup text for map. Describes site.}
   \item{popup_score}{Popup text for map. Describes parameter and score.}
   \item{alt}{Alt text for map}
 }
}
\usage{
df_score
}
\description{
Dataframe containing annual parameter scores.
}
\keyword{datasets}
