% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_thresholds}
\alias{df_thresholds}
\title{Custom thresholds (dataframe)}
\format{
A dataframe with 13 columns:
 \describe{
   \item{State}{State}
   \item{Group}{Group}
   \item{Site}{Site ID}
   \item{Depth}{Depth category}
   \item{Parameter}{Parameter}
   \item{Unit}{Parameter unit}
   \item{Calculation}{How to calculate annual score. Values: min, max, mean,
   geomean, median, 90p}
   \item{Min}{Minimum acceptable value}
   \item{Max}{Maximum acceptable value}
   \item{Excellent}{Threshold between excellent and good data}
   \item{Good}{Threshold between good and fair data}
   \item{Fair}{Threshold between fair and poor data}
   \item{Best}{Whether excellent data is higher or lower than fair data}
 }
}
\usage{
df_thresholds
}
\description{
Dataframe containing custom threshold values.
}
\keyword{datasets}
